🚚 Move: move editProfile and TravelInfo
features와 entities폴더 삭제, editProfile과 TravelInfo폴더를 widgets폴더로 변경
