+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "prepare": "husky && husky install",
+    "prepare": "husky",
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
