+++ src/app/myPage/page.tsx
@@ -1,6 +1,4 @@
-'use client';
-
-import EditProfileForm from '@/features/editProfile/ui/EditProfileForm';
+import EditProfileForm from '@/widgets/editProfile/ui/EditProfileForm';
 
 export default function MyProfilePage() {
   return <EditProfileForm />;
