+++ src/widgets/TravelInfo/components/DetailSection.tsx
