+++ src/widgets/TravelInfo/components/Map.tsx
