+++ src/widgets/TravelInfo/components/PlacesList.tsx
